trying to make softap work again
