+++ src/deployable_example.cpp
@@ -132,7 +132,17 @@ camera_fb_t *frame = NULL;
 int *last_frame_buffer = NULL;
 char groundlight_endpoint[60] = "api.groundlight.ai";
 
-Preferences preferences;
+// Preferences preferences;
+#ifdef ENABLE_AP
+  // #include <AsyncTCP.h>
+  // #include <ESPAsyncWebServer.h>
+
+  // AsyncWebServer server(80);
+
+  #include "ap_configuration.h"
+#else
+  Preferences preferences;
+#endif
 
 const bool SHOW_LOGS = false;
 void debug(String message) {
@@ -250,115 +260,6 @@ void deep_sleep() {
   esp_deep_sleep_start();
 }
 
-#ifdef ENABLE_AP
-  #include <AsyncTCP.h>
-  #include <ESPAsyncWebServer.h>
-
-  AsyncWebServer server(80);
-
-const char index_html[] PROGMEM = R"rawliteral(
-<!DOCTYPE HTML><html><head>
-  <title>ESP Input Form</title>
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <style>
-    form {
-      display: flex;
-      flex-direction: column;
-      place-items: center normal;
-    }
-    input {
-      width: 80vw;
-      margin: 10px;
-    }
-    input[type=submit] {
-      width: 100px;
-      margin: 10px;
-      background-color: #4CAF50;
-      color: white;
-      padding: 14px 20px;
-      border: none;
-      border-radius: 4px;
-      cursor: pointer;
-    }
-    input[type=submit]:hover {
-      background-color: #45a049;
-    }
-  </style>
-  </head><body>
-  <form action="/config">
-    WiFi SSID: <input type="text" name="ssid" value="%ssid%">
-    WiFi Password: <input type="text" name="pw" value="%password%">
-    Detector Id: <input type="text" name="det_id" value="%det_id%">
-    API Key: <input type="text" name="api_key" value="%api_key%">
-    Query Delay (seconds): <input type="text" name="query_delay" value="%query_delay%">
-    Endpoint: <input type="text" name="endpoint" value="%endpoint%">
-    Target Confidence: <input type="text" name="tConf" value="%tConf%">
-    Motion Alpha (float between 0 and 1): <input type="text" name="mot_a" value="%mot_a%">
-    Motion Beta (float between 0 and 1): <input type="text" name="mot_b" value="%mot_b%">
-    Stacklight UUID: <input type="text" name="sl_uuid" value="%sl_uuid%">
-    Slack URL: <input type="text" name="slack_url" value="%slack_url%">
-    Email: <input type="text" name="email" value="%email%">
-    Email Endpoint: <input type="text" name="email_endpoint" value="%email_endpoint%">
-    Email Key: <input type="text" name="email_key" value="%email_key%">
-    Email Host: <input type="text" name="email_host" value="%email_host%">
-    Twilio SID: <input type="text" name="twilio_sid" value="%twilio_sid%">
-    Twilio Token: <input type="text" name="twilio_token" value="%twilio_token%">
-    Twilio Number: <input type="text" name="twilio_number" value="%twilio_number%">
-    Twilio Recipient: <input type="text" name="twilio_recipient" value="%twilio_recipient%">
-    <input type="submit" value="Submit">
-  </form>
-</body></html>
-)rawliteral";
-
-const char sent_html[] PROGMEM = R"rawliteral(
-<!DOCTYPE HTML><html><head>
-  <title>ESP Input Form</title>
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <style>
-    h1 {
-      font-size: 2em;
-    }
-  </style>
-  </head><body>
-  <h1>Configuration Sent!</h1>
-  <a href="/">Return to Home Page</a>
-</body></html>
-)rawliteral";
-
-void notFound(AsyncWebServerRequest *request) {
-  request->send(404, "text/plain", "Not found");
-}
-
-String processor(const String& var) {
-  preferences.begin("config");
-  // String out = String();
-  String out = "";
-  // if (var == "ssid") return String(ssid);
-  if (var == "ssid") out = ssid;
-  else if (var == "password") out = password;
-  else if (var == "det_id") out = groundlight_det_id;
-  else if (var == "api_key") out = groundlight_API_key;
-  else if (var == "query_delay") out = String(query_delay);
-  else if (var == "endpoint") out = groundlight_endpoint;
-  else if (var == "tConf") out = String(targetConfidence);
-  else if (var == "mot_a" && preferences.isKey("mot_a")) out = String(preferences.getString("mot_a", "0.0"));
-  else if (var == "mot_b" && preferences.isKey("mot_b")) out = String(preferences.getString("mot_b", "0.0"));
-  else if (var == "sl_uuid" && preferences.isKey("sl_uuid")) out = preferences.getString("sl_uuid", "");
-  else if (var == "slack_url" && preferences.isKey("slack_url")) out = preferences.getString("slack_url", "");
-  else if (var == "email" && preferences.isKey("email")) out = preferences.getString("email", "");
-  else if (var == "email_endpoint" && preferences.isKey("emailEndpoint")) out = preferences.getString("emailEndpoint", "");
-  else if (var == "email_key" && preferences.isKey("emailKey")) out = preferences.getString("emailKey", "");
-  else if (var == "email_host" && preferences.isKey("emailHost")) out = preferences.getString("emailHost", "");
-  else if (var == "twilio_sid" && preferences.isKey("twilioSID")) out = preferences.getString("twilioSID", "");
-  else if (var == "twilio_token" && preferences.isKey("twilioKey")) out = preferences.getString("twilioKey", "");
-  else if (var == "twilio_number" && preferences.isKey("twilioNumber")) out = preferences.getString("twilioNumber", "");
-  else if (var == "twilio_recipient" && preferences.isKey("twilioEndpoint")) out = preferences.getString("twilioEndpoint", "");
-  preferences.end();
-  return out;
-  // return var;
-}
-#endif
-
 void setup() {
 
 #if defined(GPIO_LED_FLASH)
@@ -412,90 +313,8 @@ void setup() {
   mac.replace(":", "");
   mac = mac.substring(6);
   preferences.begin("config", false);
-  if (preferences.isKey("ap_pw")) {
-    WiFi.softAP((StringSumHelper) "ESP32-CAM-" + mac, preferences.getString("ap_pw", "").c_str());
-  } else {
-    WiFi.softAP((StringSumHelper) "ESP32-CAM-" + mac);
-  }
+  ap_setup((StringSumHelper) "ESP32-CAM-" + mac, preferences.getString("ap_pw", ""));
   preferences.end();
-  // Send web page with input fields to client
-  // at http://192.168.4.1/
-  server.on("/", HTTP_GET, [](AsyncWebServerRequest *request){
-    request->send_P(200, "text/html", index_html, processor);
-  });
-
-  server.on("/config", HTTP_GET, [] (AsyncWebServerRequest *request) {
-    preferences.begin("config", false);
-    if (request->hasParam("ssid") && request->getParam("ssid")->value() != "") {
-      preferences.putString("ssid", request->getParam("ssid")->value());
-      strcpy(ssid, request->getParam("ssid")->value().c_str());
-    }
-    if (request->hasParam("pw") && request->getParam("pw")->value() != "") {
-      preferences.putString("password", request->getParam("pw")->value());
-      strcpy(password, request->getParam("pw")->value().c_str());
-    }
-    if (request->hasParam("det_id") && request->getParam("det_id")->value() != "") {
-      preferences.putString("det_id", request->getParam("det_id")->value());
-      strcpy(groundlight_det_id, request->getParam("det_id")->value().c_str());
-    }
-    if (request->hasParam("api_key") && request->getParam("api_key")->value() != "") {
-      preferences.putString("api_key", request->getParam("api_key")->value());
-      strcpy(groundlight_API_key, request->getParam("api_key")->value().c_str());
-    }
-    if (request->hasParam("query_delay") && request->getParam("query_delay")->value() != "") {
-      query_delay = request->getParam("query_delay")->value().toInt();
-      preferences.putInt("query_delay", query_delay);
-    }
-    if (request->hasParam("endpoint") && request->getParam("endpoint")->value() != "") {
-      preferences.putString("endpoint", request->getParam("endpoint")->value());
-      strcpy(groundlight_endpoint, request->getParam("endpoint")->value().c_str());
-    }
-    if (request->hasParam("tConf") && request->getParam("tConf")->value() != "") {
-      targetConfidence = request->getParam("tConf")->value().toFloat();
-      preferences.putFloat("tConf", targetConfidence);
-    }
-    if (request->hasParam("mot_a") && request->getParam("mot_a")->value() != "") {
-      preferences.putString("mot_a", request->getParam("mot_a")->value());
-    }
-    if (request->hasParam("mot_b") && request->getParam("mot_b")->value() != "") {
-      preferences.putString("mot_b", request->getParam("mot_b")->value());
-    }
-    if (request->hasParam("sl_uuid") && request->getParam("sl_uuid")->value() != "") {
-      preferences.putString("sl_uuid", request->getParam("sl_uuid")->value());
-    }
-    if (request->hasParam("slack_url") && request->getParam("slack_url")->value() != "") {
-      preferences.putString("slack_url", request->getParam("slack_url")->value());
-    }
-    if (request->hasParam("email") && request->getParam("email")->value() != "") {
-      preferences.putString("email", request->getParam("email")->value());
-    }
-    if (request->hasParam("email_endpoint") && request->getParam("email_endpoint")->value() != "") {
-      preferences.putString("emailEndpoint", request->getParam("email_endpoint")->value());
-    }
-    if (request->hasParam("email_key") && request->getParam("email_key")->value() != "") {
-      preferences.putString("emailKey", request->getParam("email_key")->value());
-    }
-    if (request->hasParam("email_host") && request->getParam("email_host")->value() != "") {
-      preferences.putString("emailHost", request->getParam("email_host")->value());
-    }
-    if (request->hasParam("twilio_sid") && request->getParam("twilio_sid")->value() != "") {
-      preferences.putString("twilioSID", request->getParam("twilio_sid")->value());
-    }
-    if (request->hasParam("twilio_token") && request->getParam("twilio_token")->value() != "") {
-      preferences.putString("twilioKey", request->getParam("twilio_token")->value());
-    }
-    if (request->hasParam("twilio_number") && request->getParam("twilio_number")->value() != "") {
-      preferences.putString("twilioNumber", request->getParam("twilio_number")->value());
-    }
-    if (request->hasParam("twilio_recipient") && request->getParam("twilio_recipient")->value() != "") {
-      preferences.putString("twilioEndpoint", request->getParam("twilio_recipient")->value());
-    }
-    // request->send(200, "text/html", "Configuration sent to your ESP Camera<br><a href=\"/\">Return to Home Page</a>");
-    request->send_P(200, "text/html", sent_html);
-    preferences.end();
-  });
-  server.onNotFound(notFound);
-  server.begin();
 #endif
   preferences.begin("config", false);
   if (preferences.isKey("ssid") && preferences.isKey("password") && preferences.isKey("api_key") && preferences.isKey("det_id") && preferences.isKey("query_delay")) {
